Counting For
Support for counting for loops in the format
for cv = sv to endValue by incr:
    body
endfor;

sv, envValue, and incr can all be expressions
changing sv value in body will not adjust number of executions, changing endValue will not impact the number of executions. Increment must be a positive number.